chore: refactor sqlite
