+++ internal/store/sqliteDB/user.go
@@ -1,16 +1,14 @@
-package store
+package sqlitedb
 
 import (
 	"context"
 	"fmt"
-	"log"
-	"os"
 	"time"
 
 	"crawshaw.io/sqlite"
 	"crawshaw.io/sqlite/sqlitex"
-	"github.com/Milad75Rasouli/portfolio/internal/db"
 	"github.com/Milad75Rasouli/portfolio/internal/model"
+	"github.com/Milad75Rasouli/portfolio/internal/store"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 )
@@ -27,30 +25,6 @@ func NewUserSqlite(dbPool *sqlitex.Pool, logger *zap.Logger) *UserSqlite {
 		dbPool: dbPool,
 		logger: logger,
 	}
-}
-func CreateSqliteTable(dbPool *sqlitex.Pool, cfg db.Config) error {
-	ctx, cancel := context.WithTimeout(context.Background(), cfg.ConnectionTimeout)
-	defer cancel()
-	conn := dbPool.Get(ctx)
-	defer dbPool.Put(conn)
-	stmt, err := conn.Prepare(`CREATE TABLE IF NOT EXISTS user (
-		id INTEGER,
-		full_name TEXT NOT NULL,
-		email INTEGER NOT NULL UNIQUE,
-		password TEXT NOT NULL,
-		is_github INTEGER DEFAULT 0,
-		online_at DATETIME,
-		created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-		modified_at DATETIME,
-		PRIMARY KEY(id)
-		);`)
-	if err != nil {
-		return errors.Wrap(err, CannotCreateTableError.Error())
-	}
-	defer stmt.Finalize()
-	_, err = stmt.Step()
-
-	return err
 }
 func (u UserSqlite) parseToUser(stmt *sqlite.Stmt) (model.User, error) {
 	var (
@@ -103,7 +77,7 @@ func (u *UserSqlite) Create(ctx context.Context, usr model.User) (int64, error) 
 	if err != nil {
 		e := err.Error()[18:42]
 		if e == "SQLITE_CONSTRAINT_UNIQUE" {
-			return rowID, DuplicateUserError
+			return rowID, store.DuplicateUserError
 		}
 	}
 	hasRow, err := stmtSelect.Step()
@@ -130,7 +104,7 @@ func (u *UserSqlite) GetByEmail(ctx context.Context, email string) (model.User, 
 	var hasRow bool
 	hasRow, err = stmt.Step()
 	if hasRow == false {
-		return usr, UserNotFountError
+		return usr, store.UserNotFountError
 	}
 	if err != nil {
 		return usr, err
@@ -152,7 +126,7 @@ func (u *UserSqlite) GetByID(ctx context.Context, id int64) (model.User, error) 
 	var hasRow bool
 	hasRow, err = stmt.Step()
 	if hasRow == false {
-		return usr, UserNotFountError
+		return usr, store.UserNotFountError
 	}
 	if err != nil {
 		return usr, err
@@ -224,45 +198,7 @@ func (u *UserSqlite) UpdatePasswordFullName(ctx context.Context, id int64, passw
 	var hasRow bool
 	hasRow, err = stmt.Step()
 	if hasRow {
-		return UserNotFountError
+		return store.UserNotFountError
 	}
 	return err
-}
-
-type SqliteInit struct {
-	Folder string
-}
-
-func (d *SqliteInit) Init(isTestMode bool, config db.Config, logger *zap.Logger) (*UserSqlite, func(), error) {
-	var userDB *UserSqlite
-
-	os.Mkdir(d.Folder, 0777)
-	cfg := config
-	if isTestMode == true {
-		cfg = db.Config{
-			IsSqlite:          true,
-			ConnectionTimeout: time.Millisecond * 200,
-		}
-	}
-	dbPool, err := db.New(cfg)
-	if err != nil {
-		return nil, nil, err
-	}
-	err = CreateSqliteTable(dbPool, cfg)
-	if err != nil {
-		return nil, nil, err
-	}
-	userDB = NewUserSqlite(dbPool, logger)
-	return userDB, func() {
-		err := dbPool.Close()
-		if err != nil {
-			log.Printf("Error closing database connection: %v", err)
-		}
-		if isTestMode == true {
-			err = os.RemoveAll(d.Folder)
-			if err != nil {
-				log.Printf("Error removing database folder: %v", err)
-			}
-		}
-	}, nil
 }
