+++ internal/store/sqliteDB/user_test.go
@@ -1,4 +1,4 @@
-package store
+package sqlitedb
 
 import (
 	"context"
